eopkg/archive: Implement opening of .eopkg files
In essence, an .eopkg is a simple zip file. It contains a couple of
files of interest from a metadata perspective, and inside the zip we
also have an install.tar.xz containing the actual files (highly
compressed, -6).

With this in mind we're able to take advantage of Go's built in
archive/zip library to open them natively, where we can then set
about deserialising the XML metadata.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>