eopkg/package: Add utility method for finding a zip file
Mostly we are interested in finding only the history/files.xml files,
so we now provide a utility method for finding the file inside of the
archive.

We also now begin the skeleton methods for parsing the files.xml and
metadata.xml files, which will be used to fully understand the package.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>