manager: Add skeleton for the Pool implementation
The Pool will be used to track assets shared between the repositories,
allowing them to deduplicate the storage. A key advantage of this is
obviously saving on disk space, but it also allows the package data
to be deduplicated between the repositories too, paving the way
for a "pull" system.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>