server: Restrict permissions of the socket
This ensures that only the user that started the daemon can interact
with the socket, which is currently going to be the root user. We'll
actually alter this in future to make it user and primary group only,
so that users can join into the relevant group to not require any
use of sudo to interact with the local daemon, and that it can run
as an unprivileged process.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>