ferry: Start a new subpackage for Slip management
In the ferryd architecture, packages are ferried to the Slip, which is
responsible for incoming packages, organisation, etc. In the grand scheme
of things, the slip provides the core functionality of the ferryd.

This component will be responsible for all direct-disk operations with
the database and filesystem for repository management, and accepting
builder uploads & manifests ("ferried packages").

Signed-off-by: Ikey Doherty <ikey@solus-project.com>