slip: Add a new Component interface for easy management
Instead of having lots of intertwined code for managing specifics of the
startup/teardown routines, we add a new generic component interface which
allows the implementations to take care of their own business.

Currently this allows us to initialise the pool with it's mandatory root
level bucket and pool directory on disk.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>