libeopkg: Add new function for getting the source path component
To help organisation within the repository, and indeed to help with the
time it takes to list a directory (such as through a web browser), we split
the repository up into several layers. The first level is defined by either
the first letter of the source name, or if the source name is more than
4 characters and begins with "lib", we'll use those 4 characters as the
path component.

The final part of the component is the source name of the package itself,
which leads to automatic grouping of eopkgs by their logical association.
Unlike binman, ferryd will mandate lower case directories, as we've seen
some issues with "N/" "n/" paths when converting old packages from the
legacy pspec format with pristine source name, to ypkg where it is lower
case due to forming the base name of all resulting eopkg files.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>