slip: Add a new reusable gob transcoder system
Much of the database interaction is going to require back and forth with
serialisation, and will have a significant enough overhead from lazy
initialisation of new encoders/decoders.

This approach creates a per-component/routine GobTranscoder, which will
protect the writer/reader portions in a mutex, and try to reuse buffers
as much as possible. It also adds very simple wrapper methods that only
take pointers, to simplify the gob usage in ferryd.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>