slip: Add basic serialisation logic to the pool
The basic premise of the Pool is refcounting IDs, in effect its like a
disk-backed hashmap. Once we plug other parts of the chain in, we'll be
able to ref an object into the pool, meaning both the DB and the disk,
and when unreffing we'll simply remove from the pool when we hit 0,
ensuring we clean up after ourselves.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>