index: Issue repo index updates atomically
A major design flaw with the old binman implementation was rewriting the
eopkg index in place. This would always leave the potential for users to
update their repo index copy *during* this write, leading to corruption
or indeed seeing an incomplete index.

This has the effect that the repo had to be allowed to settle before we
could trust new items had truly shown up, and in quite a few cases, would
actually cause builds to fail due to the corrupted index.

To mitigate this, we write the index out to a special ".new" file which is
never pulled by package managers. Once the update is complete, we swap out
the file itself without destroying the active file descriptors. Thus if
the index has already been fetched and the file descriptor is open, the
client still sees the old state, *not* a corrupted index. We then immediately
rename the temporary file to the *new* name, so future requests will open
this new file only.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>