index: Write the oh-so-important sha1sum file
This is required by eopkg to perform update detections, and prevent the
need to constantly download the index. If the hash changes, then the client
will download the new repo index.

We now make the Index function a bit more intelligent by managing multiple
paths, and give it a way to back out of a broken index operation by tracking
the atomic file updates + new file creations.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>