Add boilerplate for the new ferry JobProcessor
Many of the jobs in ferry can only performed sequentially, because we're
actually bound by the database locking. This includes indexing the repository
and actually including the eopkgs.

However, there is scope for parallel background jobs, chiefly the production
of delta packages. These can take a significant amount of time and should
most certainly be run in parallel. However, when these are built we want
them to go back into the sequential queue for re-processing and inclusion
into the index.

Effectively we create a funneling system with a sideband for long running
operations while retaining atomic sanity on the database itself.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>