Implement the Transit Manifest format
This file is always uploaded by the builder after the eopkg files, in a
reliable order. The manifest describes the accompanying payload, and will
allow verification of the contents before they make it into the repository.

Most importantly, this ensures that we only ever attempt to process a complete
and error-free upload, i.e. no data corruption.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>