jobs: Add basic dependency management and delayed execution
In essence we want to keep things going as quickly as we can, so we'll
add job dependencies to the mix. The child and parent job will keep updating
their state independently, and a job with dependencies just shouldn't be
added for execution.

This is super primitive, but will allow us to push a job "later" once we've
dusted it out with a complete set of dependencies. Our main goal here is to
permit the construction of deltas on the main background pool, and add a new
indexing job on the sequential priority queue once the deltas have all been
completed.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>