libeopkg: Begin working on the DeltaProducer
The job of the DeltaProducer is to examine two eopkgs, and determine the
files that should remain in the *new* package. Once fully implemented, this
will basically form a list of "keepers" and copy them from the input install
file to the output install file, and will eventually then wrap it up in a new
eopkg file with the .delta.eopkg suffix.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>