libeopkg: Simplify archive interface, add rudimentary delta identification
The logic is effectively the same as how it's handled right now inside
eopkg, any hash-mismatched files are placed into the .delta.eopkg. In
the future we'd like a deduplicated format where a hash can only exist
once but be referenced multiple times.

For now we build an in-memory hit-map for files in the new package that
we'd like to keep to then put into the delta (not yet producing this though)

Signed-off-by: Ikey Doherty <ikey@solus-project.com>