libeopkg: Make the DeltaProducer slightly lighter
We don't actually need an ArchiveReader for the old eopkg, as we're purely
interested in the central record files.xml index for comparison sake. Thus
we can load a straight-up libeopkg.Package and iterate the files.xml, only
using ArchiveReader on the *new* eopkg, where we selectively copy the
records we're interested in.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>