Stop using the xzed library for deltas
Unfortunately using liblzma in-process leads to absolutely staggering
memory leaks, quickly leading us to have 30+GB leaks and causing OOM
situations.

To remedy this, we do things the long way round. Effectively we write
our the new packages install.tar.xz to disk, invoke `unxz` to decompress
it, and then copy the relevant pieces into a new install.tar file. When
done we use XzFile to invoke `xz` binary to re-compress this, and then
we'll write this back into the zip file (.delta.eopkg)

It's a bit inefficient, but for a long running process, we basically have
no choice here.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>