jobs: Begin work on the new Worker implementation
The Worker is effectively our thread type, which will know ahead of time
whether it is a sequential or parallel queue type. These will all run in
the background, specialising for a given task queue.

Additionally we'll use an incrementing sleeper logic so that we don't
needlessly spin the system waiting for jobs, and then reset the time index
once we're back in the land of fetchable jobs.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>