jobs: Restore the Delta operation to create deltas on a name
This job is quite simple and will just batch-attempt the creation of a
group of deltas for a given package name. Note that it may attempt to
create multiple deltas under one thread. Whilst not exactly ideal, this
makes sense when you have multiple packages in a single payload, and
payloads hitting the repo all the time.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>