jobs: When we finish processing tram upload, fire off delta job
To save having absurdly complex dependency trees, we'll just schedule
a delta job for each new eopkg arriving in a transit manifest upload.
It might be the case that no delta is possible, but to save duplicating
code, we'll let the DeltaJobHandler work that one out for itself.

This will ensure that every package being inserted into the repository
from builder uploads will always be evaluated for re-deltaing, which is
something that binman failed to do repeatedly.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>