jobs: Implement the mass-delta-repo operation
As a design decision to retain simplicity, this call will not directly
cause the repository to be indexed. Typically, this function is only
used after a mass import of an existing repository without deltas.
Thus, we'll add a new DeltaIndex job to ensure deltas are indexed on
the transit manifest queue, and keep this mass-delta free of any kind
of forced indexing.

This also ensures we don't incur a massive import penalty when importing
thousands of existing packages and creating thousands of additional deltas.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>