core: Expose a delta staging tree alongside delta build tree
The delta staging tree is where freshly produced delta packages will
be placed, after leaving the initial delta build area. This means that
the delta is healthy and ready for inclusion. That delta package should
then be scheduled for import into the repository along with deleting
the staged asset once it has been completed.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>