core: Stuff a DeltaInformation specifier in delta pool entries
This adds support to the pool for .delta.eopkg files, which simply
just need to retain the FromRelease and ToRelease fields, given that
a delta must be created and valid already.

This approach also allows the delta files to benefit from proper ref
counting and garbage collection down the line, to enable sharing them
across multiple repos (i.e. pulling)

Signed-off-by: Ikey Doherty <ikey@solus-project.com>