Hook up a new IncludeDelta job to chain through inclusion of .delta.eopkg
This leverages all parts of the API and will utilise the manager to set
up a locked-update-view of the pool + repo before trying to insert, ref
count, etc. The eventual goal is the entrance of the .delta.eopkg back
into the target repository, with the metadata stored within the target
repo package *name*'s entry.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>