core: Teach index how to emit DeltaPackages for the index
Quite simply we'll look at our Deltas list within the RepoEntry, and
for each one, determine if it can be applied to this "tip" package. If
the ToRelease is the same as the current tip release, then we're allow
to show this delta to users. The delta is expose with a FromRelease so
that the eopkg client knows if it can use the delta package.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>