core: Add special protection for the Index job with its own lock
The index may only be emitted in a linear fashion by one job at a time,
as we're dealing with the emission of files in a unique namespace.
This change adds a single mutex for the Index function in the cached
repository object, ensuring any subsequent call to Index will block until
the first index is complete, ensuring an atomic operation with a singular
repository state.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>