jobs: Stop spawning sync jobs from async and hitting deadlock
We've seen deadlocks when we try to go and hit a sync off an async,
so to alleviate this we're doing that target work within our async worker.
We've also modified the repos to contain their own mutex so that we lock
internally in the right places, and not get pooched by boltdb itself.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>