libdb: Implement 'Update' view, using leveldb.Batch
We'll create a temporary clone of ourselves and insert a batch operation
into this clone, so that any write-centric calls will only use the batch.
If the function passed does not error, we'll apply the batch in and return
the leveldb error. If it does fail, we'll return that error and drop the
entire batch.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>