libdb: Return a new libdb.DatabaseConnection for an opaque Database
This connection is refcounted per issue #5, to ensure we actually free
up resources when we're done with them. When the refcount hits 0 once
again, we can deallocate the underlying storage mechanism. When it hits
higher than 0 for the first time, we'll try to open the underlying storage.
This will help alleviate some the of the pains of the memdb nodes not being
reclaimed by leveldb.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>