Add new (not yet functioning) RefDelta function
This will be used to include a new delta into a repository that was initially
created for another repository, to speed up delta inclusion. That is, we might
be pulling from one repo to another, and instead of *creating* a new delta
which is expensive, we'll reuse the existing one if available.

This is currently not functional as we'll need to change the internals for
the repo object to understand how to refbump an existing package, instead
of needing a libeopkg.Package object.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>