core: Teach repo + pool how to clone an existing delta
We'll now try to reuse an existing delta package if its available, instead
of attempting the creation of new ones which would waste precious time.
Basically we extract the pool path for it, clone it in and set the delta
fields as usual, but just copy direct from the pool and increase the refcount
there.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>