core: Teach repo how to bump-ref an existing package into it
This is in line with the work we've done already for the deltas,
and allows us to clone/dupe an existing package into the target repository.
This is essential work which will allow us to support clone/pull operations
easily through the simple refcount system, and save a whole lot of time
in doing so.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>