core: Implement the 'pull' operation
When pulling - we simply check for missing or mismatched packages to pull
in. That is to say, if target has a package we don't have, include it.
If target and source both have a package by a given name, but the target
has an older release than the source, also pull that in. As clone/pull
should be used *responsibly*, we'll return an error if target has a package
that is *newer* than source. Because that's just jank.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>