Fully implement "remove source" functionality
Remove source will cause an unref cycle to begin for all pkgIDs found
to correlate to a given "source & release" record match. Internally, this
results in a sequential function call to UnrefPackage.

The new UnrefPackage function has to wear multiple hats. It is quite
possible when removing a given pkgID that we still have another candidate
package available for it. Thus, we find the new match and assign it to the
entry after having removed the on disk version.

Additionally, we may now have invalidated delta packages that either point
to or from the removed pkgID's release ID. In this instance we remove all
of those deltas as part of our package garbage collection cycle.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>