jobs: Implement record rotation in the stored jobs buckets
After a lengthy discussion with @DataDrake we reached a compromise on
the best approach to rotation of the jobs, in a write-pointer-only
circular buffer.

Basically we'll track a key to store the next key, which is wrapped to
a maximum of 100. It preserves the leveldb/libdb sequencing, but ensures
we can wrap back around to the start and overwrite old records.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>