core: Add anal mode to control what we let in. Yep. I just said this.
We must be very specific about transit manifest uploads, as this data
comes from the "every day update". If our local package is newer than
the incoming package, we need to reject that (because that's just wrong.)
as the packager has likely made an error.

We must also ensure that the release number has been bumped, it is a fatal
error to include a package with the same release number as the one that
already exists, and would break the indexing operations.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>